requirements.txt
streamlit==1.43.0
python-dotenv==1.0.0
pandas==2.0.3
requests==2.31.0
olefile==0.46
langchain==0.3.19
langchain-core==0.3.35
langchain-text-splitters==0.3.6
google-generativeai==0.3.1
google-ai-generativelanguage==0.4.0
numpy==1.26.3
pillow==10.2.0
tqdm==4.66.2
matplotlib==3.9.0
scikit-learn==1.5.0
diskcache==5.6.3
reportlab==4.3.1
mistralai==0.1.5
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
# Windows-specific dependencies with platform check
pywin32==308; platform_system=="Windows"
pyhwpx==0.44.1; platform_system=="Windows"
# Non-Windows dependencies
pyhwp==0.1b9; platform_system!="Windows" 